Make the event a public interface Move the app to its own package.
