+++ app/src/main/java/com/navin/flintstones/rxwebsocket_app/GsonRequestConvertor.java
@@ -1,7 +1,8 @@
-package com.navin.flintstones.rxwebsocket;
+package com.navin.flintstones.rxwebsocket_app;
 
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
+import com.navin.flintstones.rxwebsocket.WebSocketConverter;
 
 import java.io.IOException;
 import java.nio.charset.Charset;
