+++ app/src/main/java/com/navin/flintstones/rxwebsocket_app/GsonResponseConvertor.java
@@ -1,8 +1,9 @@
-package com.navin.flintstones.rxwebsocket;
+package com.navin.flintstones.rxwebsocket_app;
 
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
+import com.navin.flintstones.rxwebsocket.WebSocketConverter;
 
 import java.io.IOException;
 import java.io.StringReader;
