+++ app/src/main/java/com/navin/flintstones/rxwebsocket_app/WebSocketConverterFactory.java
@@ -1,8 +1,9 @@
-package com.navin.flintstones.rxwebsocket;
+package com.navin.flintstones.rxwebsocket_app;
 
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
 import com.google.gson.reflect.TypeToken;
+import com.navin.flintstones.rxwebsocket.WebSocketConverter;
 
 import java.lang.reflect.Type;
